Create README.md
this is a repository for adding bible verses to a magic mirror module. you can paste these lines including the quotation marks and coma into the compliments module on the default "compliments.js" found in the /modules/compliments/compliments folder in your magic mirror folder. just past the verses you like or all of them into the sections labeled "anytime" "morning" "afternoon" or "evening"... bible verses at certain points throughout the day on your mirror are a sure way to bring inspiration while you check your hair, or even the weather.. all thanks to your magic mirror!
here is the source of the books i am putting into the module: http://www.catholic-bible.org/ feel free to do any book of the bible you like but personally will be completing the wisdom books specifically except the book of job. 
to add to the project copy each chapter of bible to a text file from http://www.catholic-bible.org/ . then my preferred method of adding quotations is by using notepad++ found here for windows> Notepad++ v7.9.5 bug-fixes:
 make sure you go through and delete any empty lines (every other line). you can add quotations to each line by opening the file in notepad++ and click "search>replace" once the replace window comes up to add a quotation mark to the begining of each line do the following: in the find what bar type ^ then in the replace with bar type " make sure "regular expression" is selected in the bottom left of the box and then hit enter.
next to add ", to the end of each line- find what type $ replace with type ", make sure regular expression in the bottom corner is selected you may have to delete the last comma on the final line of the chapter or book you are contributing for the module to work correctly.
feel free to add an entire book. please only use the bible given in the read me for this project so that everything canonically accurate. 

56 readme readme tjsteed13